requirements.txt
pbr>=0.5.21,<1.0

eventlet>=0.13.0
Flask>=0.10,<1.0a0
iso8601>=0.1.4
kombu>=2.4.8
oslo.config>=1.2.0
python-novaclient>=2.15.0
netaddr
python-keystoneclient>=0.3.2
Routes>=1.12.3
SQLAlchemy>=0.7.8,<=0.7.99
stevedore>=0.10
WebOb>=1.2.3,<1.3a0
